add Ball class

+++ Source/Paddle.cpp
@@ -7,7 +7,7 @@
 #include "Game.h"
 #include "InputHandler.h"
 
-Paddle::Paddle(const PaddleType paddle_type, const Vec2<int> position, const Vec2<int> size, const Colour colour)
+Paddle::Paddle(const PaddleType paddle_type, const Vec2<int> position, const Vec2<unsigned int> size, const Colour colour)
 {
     this->paddle_type = paddle_type;
     this->position = position;
@@ -15,7 +15,7 @@ Paddle::Paddle(const PaddleType paddle_type, const Vec2<int> position, const Vec
     this->colour = colour;
 }
 
-void Paddle::init(const PaddleType paddle_type, const Vec2<int> position, const Vec2<int> size, const Colour colour)
+void Paddle::init(const PaddleType paddle_type, const Vec2<int> position, const Vec2<unsigned int> size, const Colour colour)
 {
     this->paddle_type = paddle_type;
     this->position = position;
@@ -43,12 +43,12 @@ void Paddle::setPosition(const Vec2<int> position)
     this->position = position;
 }
 
-Vec2<int> Paddle::getSize()
+Vec2<unsigned int> Paddle::getSize()
 {
     return size;
 }
 
-void Paddle::setSize(const Vec2<int> size)
+void Paddle::setSize(const Vec2<unsigned int> size)
 {
     this->size = size;
 }
@@ -132,7 +132,7 @@ void Paddle::update()
 
 void Paddle::render(const float interpolation)
 {
-    SDL_Rect rect = {position.x, position.y, size.x, size.y};
+    SDL_Rect rect = {position.x, position.y, static_cast<int>(size.x), static_cast<int>(size.y)};
 
     printf("%i, %i\n", position.x, position.y);
     SDL_SetRenderDrawColor(Game::renderer, colour.r, colour.g, colour.b, colour.a);
